separate k8s operations from pai
