+++ deployment/k8sPaiLibrary/maintainlib/k8s_util.py
@@ -1,0 +1,38 @@
+# Copyright (c) Microsoft Corporation
+# All rights reserved.
+#
+# MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+# documentation files (the "Software"), to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
+# to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+# The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
+# BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+# DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+
+import os
+import logging
+import logging.config
+import importlib
+from . import clean
+from . import deploy
+from . import common as pai_common
+
+
+logger = logging.getLogger(__name__)
+
+def maintain_cluster_k8s(cluster_config, **kwargs):
+
+    if kwargs["option_name"] == "deploy":
+        job_instance = deploy.deploy(cluster_config, **kwargs)
+        job_instance.run()
+    elif kwargs["option_name"] == "clean":
+        job_instance = clean.clean(cluster_config, **kwargs)
+        job_instance.run()
+
